update: 캠프 카드 레이아웃 및 더미 데이터 적용
